Simplify id callback
